fix the key type of WordCountMapper

+++ examples/word_count_mapper.cc
@@ -7,7 +7,7 @@
 #include "word_count_mapper.h"
 
 void WordCountMapper::Map(std::unordered_map<std::string, std::vector<int>> &kvs,
-                          const int &key,
+                          const long long int &key,
                           const std::string &value) {
   size_t cur = 0, pos = 0;
   while ((pos = value.find_first_of(" ", cur)) != std::string::npos) {
